requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
httpx==0.24.0
pyodbc>=5.0.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
psycopg2-binary